+++ lib/droonga/plugins/system/statistics.rb
@@ -19,7 +19,7 @@ require "droonga/database_scanner"
 module Droonga
   module Plugins
     module System
-      class ObjectCountHandler < Droonga::Handler
+      class StatisticsObjectCountHandler < Droonga::Handler
         include DatabaseScanner
 
         def handle(message)
@@ -44,8 +44,8 @@ module Droonga
       end
 
       define_single_step do |step|
-        step.name = "system.object.count"
-        step.handler = ObjectCountHandler
+        step.name = "system.statistics.object.count"
+        step.handler = StatisticsObjectCountHandler
         step.collector = Collectors::RecursiveSum
       end
     end